+++ spec/PHPCR/Shell/Query/UpdateParserSpec.php
@@ -13,6 +13,8 @@ use PHPCR\Query\QOM\PropertyValueInterface;
 use PHPCR\Query\QOM\LiteralInterface;
 use PHPCR\Query\QOM\ComparisonInterface;
 use PHPCR\Query\QueryInterface;
+use PHPCR\Shell\Query\FunctionOperand;
+use PHPCR\Shell\Query\ColumnOperand;
 
 class UpdateParserSpec extends ObjectBehavior
 {
@@ -66,16 +68,34 @@ EOT;
 
         $res->offsetGet(0)->shouldHaveType('PHPCR\Query\QueryInterface');
         $res->offsetGet(1)->shouldReturn(array(
-            'parent.foo' => array(
+            array(
                 'selector' => 'parent',
                 'name' => 'foo',
                 'value' => 'PHPCR\\FOO\\Bar',
             ),
-            'parent.bar' => array(
+            array(
                 'selector' => 'parent',
                 'name' => 'bar',
                 'value' => 'foo',
             ),
         ));
+    }
+
+    function it_should_parse_functions (
+        QueryObjectModelFactoryInterface $qomf,
+        SourceInterface $source,
+        QueryInterface $query
+    )
+    {
+        $qomf->selector('a', 'dtl:article')->willReturn($source);
+        $qomf->createQuery($source, null)->willReturn($query);
+
+
+        $sql = <<<EOT
+UPDATE [dtl:article] AS a SET a.tags = array_replace(a.tags, 'asd', 'dsa')
+EOT;
+        $res = $this->parse($sql);
+
+        $res->offsetGet(0)->shouldHaveType('PHPCR\Query\QueryInterface');
     }
 }